cdk/requirements.txt
-e .
aws_cdk.aws_cognito
aws_cdk.cdk
